# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: POS API
  version: 1.0.0
  description: POSシステムのAPI仕様
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/item-types:
    get:
      summary: アイテムタイプ一覧取得
      operationId: getItemTypes
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ItemTypeResponse'
    post:
      summary: アイテムタイプ作成
      operationId: createItemType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemTypeCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTypeResponse'
  /api/item-types/{id}:
    get:
      summary: idからアイテムタイプ情報取得
      operationId: getItemType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTypeResponse'
    put:
      summary: アイテム情報更新
      operationId: updateItemType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemTypeUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTypeResponse'
    delete:
      summary: アイテムタイプ削除
      operationId: deleteItemType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
...
components:
  schemas:
...
    ItemTypeResponse:
      type: object
      required:
        - id
        - name
        - display_name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_name:
          type: string
    ItemTypeCreateRequest:
      type: object
      required:
        - name
        - display_name
      properties:
        name:
          type: string
        display_name:
          type: string
    # 更新リクエスト用
    ItemTypeUpdateRequest:
      type: object
      required:
        - id
        - name
        - display_name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_name:
          type: string
    # 指名を入れたItem
...